@@ -38,10 +38,8 @@ namespace DotVVM.Samples.MiniProfiler.AspNetCore
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            loggerFactory.AddConsole();
-
             app.UseMiniProfiler();
 
             app.UseDotVVM<DotvvmStartup>(env.ContentRootPath);
